gcs-bigquery-loader/app/requirements.txt
google-cloud-bigquery==1.11.2 